+++ src/bract/ring/inducer.clj
@@ -12,7 +12,7 @@
     [bract.core.inducer :as core-inducer]
     [bract.core.type    :as core-type]
     [bract.core.util    :as core-util]
-    [bract.ring.config  :as ring-config]))
+    [bract.ring.keydef  :as ring-kdef]))
 
 
 (defn apply-wrappers
@@ -23,10 +23,10 @@
   (->> wrappers
     (map (fn [wrapper-spec] (core-type/->Function
                               (fn [ctx]
-                                (let [handler (ring-config/ctx-ring-handler ctx)
+                                (let [handler (ring-kdef/ctx-ring-handler ctx)
                                       wrapper (core-type/ifunc wrapper-spec)]
                                   (->> (wrapper handler ctx)
-                                    (assoc ctx (key ring-config/ctx-ring-handler)))))
+                                    (assoc ctx (key ring-kdef/ctx-ring-handler)))))
                               (core-type/iname wrapper-spec)
                               [])))
     (core-inducer/induce context)))