+++ test/bract/ring/inducer_test.clj
@@ -10,7 +10,7 @@
 (ns bract.ring.inducer-test
   (:require
     [clojure.test :refer :all]
-    [bract.ring.config  :as ring-config]
+    [bract.ring.keydef  :as ring-kdef]
     [bract.ring.inducer :as inducer]))
 
 
@@ -34,7 +34,7 @@
     (let [context {:bract.ring/ring-handler identity}]
       (vreset! holder 0)
       (let [new-context (inducer/apply-wrappers context [wrapper-inc wrapper-add2])
-            new-handler (ring-config/ctx-ring-handler new-context)]
+            new-handler (ring-kdef/ctx-ring-handler new-context)]
         (is (contains? new-context :bract.ring/ring-handler))
         (is (= :foo (new-handler :foo))))
       (is (= 3 @holder))))
@@ -43,7 +43,7 @@
       (vreset! holder 0)
       (let [new-context (inducer/apply-wrappers context '[bract.ring.inducer-test/wrapper-inc
                                                           bract.ring.inducer-test/wrapper-add2])
-            new-handler (ring-config/ctx-ring-handler new-context)]
+            new-handler (ring-kdef/ctx-ring-handler new-context)]
         (is (contains? new-context :bract.ring/ring-handler))
         (is (= :foo (new-handler :foo))))
       (is (= 3 @holder))))